requirements.txt
apache-airflow
boto3
requests
psycopg2-binary
dbt-core
dbt-postgres
